backend/tomato_app/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
tensorflow==2.19.0
pillow
numpy==2.0.2
python-dotenv
supabase
werkzeug
google-cloud-aiplatform
google-cloud-core
google-api-core
google-genai
googleapis-common-protos
uuid
gunicorn==21.2.0